+++ mentor/src/main/kotlin/com/sourceplusplus/mentor/base/MentorJobConfig.kt
@@ -1,4 +1,4 @@
-package com.sourceplusplus.mentor
+package com.sourceplusplus.mentor.base
 
 /**
  * [MentorJob] processing configuration.
