+++ mentor/src/main/kotlin/com/sourceplusplus/mentor/base/MentorJobEvent.kt
@@ -1,4 +1,4 @@
-package com.sourceplusplus.mentor
+package com.sourceplusplus.mentor.base
 
 /**
  * Represents events [MentorJob]s may emit.
