+++ mentor/src/main/kotlin/com/sourceplusplus/mentor/base/MentorJobListener.kt
@@ -1,4 +1,4 @@
-package com.sourceplusplus.mentor
+package com.sourceplusplus.mentor.base
 
 /**
  * Listen for [MentorJobEvent]s produced by processing [MentorJob]s.
