+++ mentor/src/main/kotlin/com/sourceplusplus/mentor/impl/job/RampDetectionMentor.kt
@@ -1,13 +1,13 @@
-package com.sourceplusplus.mentor.job
+package com.sourceplusplus.mentor.impl.job
 
-import com.sourceplusplus.mentor.MentorJob
-import com.sourceplusplus.mentor.MentorTask
-import com.sourceplusplus.mentor.task.analyze.CalculateLinearRegression
-import com.sourceplusplus.mentor.task.general.DelayTask
-import com.sourceplusplus.mentor.task.monitor.GetEndpoints
-import com.sourceplusplus.mentor.task.monitor.GetService
-import com.sourceplusplus.mentor.task.monitor.GetServiceInstance
-import com.sourceplusplus.mentor.task.monitor.GetTraces
+import com.sourceplusplus.mentor.base.MentorJob
+import com.sourceplusplus.mentor.base.MentorTask
+import com.sourceplusplus.mentor.impl.task.analyze.CalculateLinearRegression
+import com.sourceplusplus.mentor.impl.task.general.DelayTask
+import com.sourceplusplus.mentor.impl.task.monitor.GetEndpoints
+import com.sourceplusplus.mentor.impl.task.monitor.GetService
+import com.sourceplusplus.mentor.impl.task.monitor.GetServiceInstance
+import com.sourceplusplus.mentor.impl.task.monitor.GetTraces
 import com.sourceplusplus.protocol.artifact.trace.TraceOrderType
 import com.sourceplusplus.protocol.portal.QueryTimeFrame
 import io.vertx.core.Vertx
