+++ mentor/src/main/kotlin/com/sourceplusplus/mentor/impl/task/analyze/DetermineThrowableLocation.kt
@@ -1,8 +1,8 @@
-package com.sourceplusplus.mentor.task.analyze
+package com.sourceplusplus.mentor.impl.task.analyze
 
-import com.sourceplusplus.mentor.MentorJob
-import com.sourceplusplus.mentor.MentorJob.ContextKey
-import com.sourceplusplus.mentor.MentorTask
+import com.sourceplusplus.mentor.base.MentorJob
+import com.sourceplusplus.mentor.base.MentorJob.ContextKey
+import com.sourceplusplus.mentor.base.MentorTask
 import com.sourceplusplus.protocol.advice.informative.ActiveExceptionAdvice
 import com.sourceplusplus.protocol.artifact.ArtifactLocation
 import com.sourceplusplus.protocol.artifact.ArtifactQualifiedName
