+++ mentor/src/main/kotlin/com/sourceplusplus/mentor/impl/task/monitor/GetServiceInstance.kt
@@ -1,8 +1,8 @@
-package com.sourceplusplus.mentor.task.monitor
+package com.sourceplusplus.mentor.impl.task.monitor
 
-import com.sourceplusplus.mentor.MentorJob
-import com.sourceplusplus.mentor.MentorJob.ContextKey
-import com.sourceplusplus.mentor.MentorTask
+import com.sourceplusplus.mentor.base.MentorJob
+import com.sourceplusplus.mentor.base.MentorJob.ContextKey
+import com.sourceplusplus.mentor.base.MentorTask
 import com.sourceplusplus.monitor.skywalking.track.ServiceInstanceTracker.Companion.getServiceInstances
 import monitor.skywalking.protocol.metadata.GetAllServicesQuery
 import monitor.skywalking.protocol.metadata.GetServiceInstancesQuery
