+++ mentor/src/test/kotlin/com/sourceplusplus/mentor/impl/task/CalculateLinearRegressionTest.kt
@@ -1,10 +1,10 @@
-package com.sourceplusplus.mentor.task
+package com.sourceplusplus.mentor.impl.task
 
-import com.sourceplusplus.mentor.MentorJob
-import com.sourceplusplus.mentor.MentorTask
+import com.sourceplusplus.mentor.base.MentorJob
+import com.sourceplusplus.mentor.base.MentorTask
 import com.sourceplusplus.mentor.MentorTest
-import com.sourceplusplus.mentor.task.analyze.CalculateLinearRegression
-import com.sourceplusplus.mentor.task.monitor.GetTraces
+import com.sourceplusplus.mentor.impl.task.analyze.CalculateLinearRegression
+import com.sourceplusplus.mentor.impl.task.monitor.GetTraces
 import com.sourceplusplus.protocol.artifact.trace.TraceOrderType
 import com.sourceplusplus.protocol.portal.QueryTimeFrame
 import io.vertx.core.Vertx
