+++ mentor/src/test/kotlin/com/sourceplusplus/mentor/impl/task/GetTracesTest.kt
@@ -1,10 +1,9 @@
-package com.sourceplusplus.mentor.task
+package com.sourceplusplus.mentor.impl.task
 
-import com.sourceplusplus.mentor.MentorJob
-import com.sourceplusplus.mentor.MentorTask
+import com.sourceplusplus.mentor.base.MentorJob
+import com.sourceplusplus.mentor.base.MentorTask
 import com.sourceplusplus.mentor.MentorTest
-import com.sourceplusplus.mentor.task.monitor.GetTraceStacks
-import com.sourceplusplus.mentor.task.monitor.GetTraces
+import com.sourceplusplus.mentor.impl.task.monitor.GetTraces
 import com.sourceplusplus.protocol.artifact.trace.TraceOrderType
 import com.sourceplusplus.protocol.portal.QueryTimeFrame
 import io.vertx.core.Vertx
@@ -13,12 +12,12 @@ import kotlinx.coroutines.runBlocking
 import org.junit.Assert.assertNotNull
 import org.junit.Test
 
-class GetTraceStacksTest : MentorTest() {
+class GetTracesTest : MentorTest() {
 
     @Test(timeout = 5000)
-    fun latestTraceStacks() {
+    fun latestTraces() {
         val emptyJob = object : MentorJob() {
-            override val vertx: Vertx = this@GetTraceStacksTest.vertx
+            override val vertx: Vertx = this@GetTracesTest.vertx
             override val tasks: List<MentorTask> = listOf()
         }
 
@@ -27,11 +26,8 @@ class GetTraceStacksTest : MentorTest() {
                 orderType = TraceOrderType.LATEST_TRACES,
                 timeFrame = QueryTimeFrame.LAST_15_MINUTES
             ).executeTask(emptyJob)
-            GetTraceStacks(
-                GetTraces.TRACE_RESULT
-            ).executeTask(emptyJob)
 
-            assertNotNull(emptyJob.context.get(GetTraceStacks.TRACE_STACKS))
+            assertNotNull(emptyJob.context.get(GetTraces.TRACE_RESULT))
         }
     }
 }
